utils: have pi_f return float

+++ src/utils/pi.hpp
@@ -8,7 +8,7 @@ constexpr double pi()
     return std::atan2(0.0, -1.0);
 }
 
-constexpr double pi_f()
+constexpr float pi_f()
 {
     return std::atan2(0.0f, -1.0f);
 }
